mcp_server/requirements.txt
# MCP Server Dependencies
fastmcp>=0.1.0
loguru>=0.7.0
python-dotenv>=1.0.0

# Shared with ingestion
torch>=2.0.0
transformers>=4.30.0
pillow>=10.0.0
azure-search-documents>=11.4.0
azure-identity>=1.15.0
azure-core>=1.29.0
